Why Cloudant? to Cloudant Basics
Why Cloudant? reads as marketing language (and likely soon be a marketing page all its own). It's definitely a question that Cloudant needs to answer, but here I think we should focus on orienting customers and would-be customers on the elements before they dive deeper into the docs or into building something with Cloudant -- JSON, HTTP, etc.